SG-17535 libubootenv: fw_setenv: Restore old behavior
This patch brings back the behavior of the U-Boot based fw_printenv
tool, allowing it to be used as a drop-in replacement for the U-Boot
tools.

Please note: The new behavior is probably better as it aligns with the
format of "env import -t". However, until SG-17574 is implemented, we
need to stick to the old format.

This commit also clarifies which patches we want to upstream and moves
them into a version-dependent directory to underline the fact that
(some) patches need to be dropped by the next update.