SG-18437 wpa-supplicant: at91sam: Avoid gateway degradation
When a gateway has neither Ethernet link nor WLAN configured, hostapd
gets started to create an AP. As wpa_supplicant and hostapd are running
mutually exclusive, the wpa_supplicant unit gets then terminated.

On the at91sam gateway, we have some long-running ExecStartPre= steps
for the Wi-Fi units, causing the wpa_supplicant unit to enter the failed
state when being terminated before ExecStart= got reached. This then
degrades the whole system.

The same can happen for the hostapd unit when being terminated early on.

Please note: The exit codes of fast-running ExecStartPre= steps are not
ignored as chances are lower for them to be affected and the wish to
know when something goes sideways prevails. Also, those steps have
proven to not be a serious source of gateway degradations (at least on
the mt7688 platform).