SG-19794 dnsmasq: start/restart/stop when hostapd does
In some situations, hostapd@wlan0.service seems to fail to start (on HC
gateways, probably due to the driver switch (SG-14607)) causing
dnsmasq@wlan0.service to fail as well. hostapd would then automatically
be restarted but not the DHCP server (dnsmasq), resulting in the clients
not receiving an IP address.