Switch to iptables-nft
The main reason is that we want to do stateless NAT of NTP packets for
gateway3 using nftables. Using iptables-nft instead of iptables-legacy
is not strictly required but integrated better when mixing nft and
iptables. For instance, you can see rules created by iptables using
`nft list ruleset`.

Reagarding why nftables in first place? It's just better in every way:
https://wiki.nftables.org/wiki-nftables/index.php/Main_differences_with_iptables

Long-term we should migrate everything else to nftables as well. That'd
mostly be things like the firewall and nngforward in meta-gardena.

Issue: SG-20159