swupdate-check: Adjust expected exit status in `systemd` unit
Previously the exit code `4` of `swupdate` was accepted as non-error
status in the `systemd` unit file of `swupdate-check`.
But the behaviour in `swupdate` changed regarding the exit status of the
downloader subprocess. Now it returns `EXIT_FAILURE` if anything goes
wrong. Previously it returned an error code that was propagated as
return value of the main process.

The commit in the `swupdate` repo that introduced the change is:
75cc133 ("start_download: return EXIT_FAILURE instead of result on failures")

Issue: SG-21041