conf: distro: Add Rust to SDK
Allow non-gateway developers to use the SDK to build their Rust based
applications.

Unfortunately, this does not build on AT91SAM. Since we are far behind the
openembedded-core master, debugging this seems like a waste of time and adding
the dependency just for the MT7688 based SDK allows to get around it.

Issue: SG-20411