rsyslog: Use port 514 for dev server
This is needed in order to fix the tests. Configuring rsyslog on the APU
to listen on port 443 will (currently) causes the boot to hang due to
the `sshtunnel-check` service and may result in other conflicts in the
future. As we do not use encryption for the dev server, let's use port
514, rsyslog's default.

Issue: SG-20937