swupdate: Downgrade to version 2020.11
`swupdate` changed the exit status of the application in case a script
embedded in the image disallows an update. We use an embedded script
to check if the update show proceed.
The change in `swupdate` is considered a fix. But we rely on the old
behavior.

Issue: SG-18399
Issue: SG-21041