ppp: Prepare for Zephyr-specific PPP service
In order to change the ppp0 baudrate for the Zephyr based RM FW, we need
a Zephyr specific ppp.service file.

This commit does not yet implement the baud rate change, but allows to
implement the change easily.

By moving the (soon to be) Zephyr specific parts from the ppp to the
(already RM FW type specific) ppp-network/ppp-network-zephyr packages, the
ppp recipe can stay agnostic.

The content of the generated packages (apart from the revisions) should
not be changed by this commit.

Issue: SG-21284