zram-init: Switch to ext4 file system
This fixes the 2038 issue for file timestamps in `/run/log/journal` (and
removes annoying log entries).

Also, ext4 seems to be a good option for use with zram as it supports
the `discard` mount option which should make sure that RAM is freed up
again when files are deleted.

Issue: SG-21150