zram-init: Decrease maximum memory usage
Adjust to the changes done in commit 15fc55e ("systemd-conf: Reduce
maximum journal size to 32MB").

We do not expect zram to use up more than ~10 MB. Though, to be on the
safe side, we set a higher value (which should still be OK).

The file system size is left as it is so that developers can manually
increase the maximum journal size if they choose to do so.

Issue: SG-21278