images: install iproute2-tc in FOSS base image
The tc command is required for the healthcheck to get interface
statistics for ppp0 in order to determine whether any packets were
dropped. It can also be useful for debugging, e.g. to show how many
packets are currently in the queue.

The tool is needed at least for all BNW(-Zephyr) images, but it is
useful in all images as we might also lose packets in the Linux
network queue with the legacy RM stack, and having tc installed allows
the healthcheck to detect this.