swupdate-legacy: Add our customizations
For the new Gateway software rollout concept, `swupdate` shall be
updated and adapted to be able to connect to an update server. To keep
the old way of updating for "legacy" images, we need a recipe with a
different name.

The files are copied (more or less) as-is from the `swupdate` recipe.

Issue: SG-21568