conf: machine: sg-mt7688: Set KERNEL_IMAGETYPE_REPLACEMENT
For some reason, we need to build the uncompressed kernel and let the
build system compress it before including it in the image.

Controlling the behavior through the configuration variable allows us to
get rid of the hack in openembedded-core, commit 34f6114d52 ("fitimage:
MIPS: Build uncompressed image") in openembedded-core.

Issue: SG-19337
Issue: SG-21192