python3: Set 2 as the default optimization level
The default optimization level of the Python interpreter for the target,
is set to 2. This is as providing the `-OO` option to the interpreter or
setting the environment variable `PYTHONOPTIMIZE=2`.

This change was rejected upstream for being to specific. Also patching
the code was seen as being suitable in this case.

Issue: SG-22583