swupdate: Set log level to "trace"
Currently, the polling interval is set to 1 minute. This causes the log
to fill up rather quickly. By lowering the log level to "trace", we
substantially reduce the number of printed lines while maintaining
visibility into when the server is contacted.

Note that only the DDI image is affected by this change.