images: Add package for support data consent settings
Currently the package installs files to the `rootfs` to indicate that
the customer allows us to collect support data (logs, metrics and
support bundles). In the future these files will not be part of the
images, but can be set by the customer (with the app). In that case wee
keep the files after a software update.

Issue: SG-22111